requirements.txt
Flask   
ultralytics
Pillow
torch
numpy
